+++ src/pylon_camera/encoding_conversions.cpp
@@ -90,6 +90,10 @@ bool ros2GenAPI(const std::string& ros_enc, std::string& gen_api_enc)
     {
         gen_api_enc = "BayerGR16";
     }
+    else if ( ros_enc == "bayer_rggb12" )
+    {
+        gen_api_enc = "BayerRG12";
+    }
     /*
     else if ( ros_enc == sensor_msgs::image_encodings::YUV422 )
     {
@@ -163,6 +167,10 @@ bool genAPI2Ros(const std::string& gen_api_enc, std::string& ros_enc)
     else if ( gen_api_enc == "BayerGR16" )
     {
         ros_enc = sensor_msgs::image_encodings::BAYER_GRBG16;
+    }
+    else if ( gen_api_enc == "BayerRG12" )
+    {
+        ros_enc = sensor_msgs::image_encodings::BAYER_RGGB16;
     }
     /*
     else if ( gen_api_enc == "YCbCr422_8" )